:recycle: [RefreshToken] Migrate
