--- config/doctrine/RefreshToken.orm.yml
@@ -1,34 +1,0 @@
-App\Entity\RefreshToken:
-    type: entity
-    table: refreshtoken
-    id:
-        id:
-            type: integer
-            id: true
-            generator:
-                strategy: AUTO
-    fields:
-        token:
-            type: string
-            length: 255
-            colomn: token
-            unique: true
-        expiresAt:
-            type: integer
-            nullable: true
-            colomn: expires_at
-        scope:
-            type: string
-            length: 255
-            colomn: scope
-            nullable: true
-    manyToOne:
-        client:
-            targetEntity: App\Entity\Attributes\Client
-            joinColumn:
-                nullable: false
-        user:
-            targetEntity: App\Entity\User
-            inversedBy: refreshTokens
-            joinColumn:
-                nullable: true
